SNBT Diff
Added a command to run the script with `node --watch`, which is essentially live reload, but just for Node scripts. Super helpful to have for debugging the output of your scripts!

Also added a script to get a diff of the prettified SNBT, as to help figure out what the RegEx calls are doing to the data. I like reverse engineering haha.