Action Traversal
Wow! That was kinda hard to debug. Figured out that it's a Trident holding other multiple entities, interestingly enough. Now it's fully diff-able!

The SNBT newline feature is actually a bug in NBTify I've been meaning to fix for some time now, and it actually helped make the SNBT more diffable, because the newlines weren't escaped in the resulting output. Oopsie, but glad it could help!

https://github.com/Offroaders123/NBTify/issues/26