Nice SNBT
Was working on updating NBTify to the latest version with the latest bug fixes and such, but then I discovered other issues with NBTify itself which prevented it from working with this project in particular. Went and fixed those, and now I'm upstreaming those changes back to here. And yes, it's a goofy version number on purpose. Because why not?? Lol.