Diff Comparison
This is a baseline output to diff my script outputs' results, running against the original RegExp-based script. I'd rather use Git to more statically analyze the outputs more safely. I was mostly using Diffchecker and `git diff` to see the differences between running the scripts, but it's hard to keep track of when something changed or not, and to see when that happens.