SNBT Types
Defined type definitions for the data that the script is working with. This helps to validate that things are running safely against the shape of the data, since things can tend to get out of sync otherwise.

Started work yesterday on a refactor of my refactor, but I didn't feel safe on changing things without validating against the shape of the SNBT content itself. Doing this will make things much faster since you don't have to copy the code line-for-line, as you will get type errors if you just use the data wrong.

Another side effect I realized in doing this is that you can kind of use this to validate that both the new version and the original are traversing the objects in the same way. But, this isn't quite the case here though, because the first one uses RegExp to edit the data, so I guess it doesn't really apply here too much. But that could definitely apply to other migrations kind of like these.