JSON Formatting + SNBT Files
Realized that the JSON for the output file doesn't need to be indented, since that's not what the original script did.

I also moved over the `.txt` file input and output extensions to `.snbt` instead, so it can be nicer to view in editors that support SNBT syntax highlighting. I found a VSCode extension which has that, so it looks really nice compared to just plain text. I want to get that working for Dovetail too! Speaking of that, I need to get standalone SNBT working for Dovetail too, since right now it doesn't have a way to open plain-text SNBT files, right now it can only edit binary NBT files *using* SNBT.